+++ src/store/middlewares/redirect.js
@@ -1,5 +1,5 @@
 import browserHistory from '../../browser-history';
-import {ActionType} from '../action';
+import {ActionType} from '../actions';
 
 export const redirect = (_store) => (next) => (action) => {
   if (action.type === ActionType.REDIRECT_TO_ROUTE) {
