+++ src/store/reducers/user/user.js
@@ -1,5 +1,5 @@
 import {AuthorizationStatus} from '../../../const';
-import {ActionType} from '../../action';
+import {ActionType} from '../../actions';
 import {extend} from '../../../utils';
 
 const initialState = {
